Refactorización Database Package
Se reestructura las clases del paquete Database para agrupar en dos
partes (acciones y sentencias SQL) los métodos de los motores de base
de datos. Se centraliza en el “engine” las clases auxiliares liberando
de este tratamiento la clase principal Database
