+++ .gitignore
@@ -12,3 +12,5 @@
 config.php
 generateDoc.sh
 plugin.list
+*.cache
+*.lock
