+++ Core/Base/DataBase.php
@@ -20,7 +20,6 @@
 namespace FacturaScripts\Core\Base;
 
 use FacturaScripts\Core\Base\DataBase\DatabaseEngine;
-use FacturaScripts\Core\Base\DataBase\DataBaseUtils;
 use FacturaScripts\Core\Base\DataBase\Mysql;
 use FacturaScripts\Core\Base\DataBase\Postgresql;
 
@@ -48,13 +47,7 @@ class DataBase
      * @var DatabaseEngine
      */
     private static $engine;
-
-    /**
-     * El enlace con las utilidades de base de datos.
-     * @var DataBaseUtils
-     */
-    private static $utils;
-
+        
     /**
      * Gestiona el log de todos los controladores, modelos y base de datos.
      * @var MiniLog
@@ -104,10 +97,6 @@ class DataBase
                     self::$miniLog->critical('No se reconoce el tipo de conexión. Debe ser MySQL o PostgreSQL');
                     break;
             }
-
-            if (self::$engine !== null) {
-                self::$utils = new DataBaseUtils(self::$engine);
-            }
         }
     }
 
@@ -150,7 +139,7 @@ class DataBase
     public function getColumns($tableName)
     {
         $result = [];
-        $aux = $this->select(self::$engine->sqlColumns($tableName));
+        $aux = $this->select(self::$engine->utilsSQL->sqlColumns($tableName));
         if (is_array($aux) && !empty($aux)) {
             foreach ($aux as $data) {
                 $result[] = self::$engine->columnFromData($data);
@@ -168,9 +157,9 @@ class DataBase
     public function getConstraints($tableName, $extended = false)
     {
         if ($extended) {
-            $sql = self::$engine->sqlConstraintsExtended($tableName);
+            $sql = self::$engine->utilsSQL->sqlConstraintsExtended($tableName);
         } else {
-            $sql = self::$engine->sqlConstraints($tableName);
+            $sql = self::$engine->utilsSQL->sqlConstraints($tableName);
         }
 
         $data = $this->select($sql);
@@ -185,7 +174,7 @@ class DataBase
     public function getIndexes($tableName)
     {
         $result = [];
-        $data = $this->select(self::$engine->sqlIndexes($tableName));
+        $data = $this->select(self::$engine->utilsSQL->sqlIndexes($tableName));
         if (is_array($data) && !empty($data)) {
             foreach ($data as $row) {
                 $result[] = ['name' => $row['Key_name']];
@@ -375,7 +364,7 @@ class DataBase
      */
     public function lastval()
     {
-        $aux = $this->select(self::$engine->sqlLastValue());
+        $aux = $this->select(self::$engine->utilsSQL->sqlLastValue());
         return $aux ? $aux[0]['num'] : false;
     }
 
@@ -432,7 +421,7 @@ class DataBase
      */
     public function generateTable($tableName, $xmlCols, $xmlCons)
     {
-        return self::$utils->generateTable($tableName, $xmlCols, $xmlCons);
+        return self::$engine->utils->generateTable($tableName, $xmlCols, $xmlCons);
     }
 
     /**
@@ -445,7 +434,7 @@ class DataBase
      */
     public function compareConstraints($tableName, $xmlCons, $dbCons, $deleteOnly = false)
     {
-        return self::$utils->compareConstraints($tableName, $xmlCons, $dbCons, $deleteOnly);
+        return self::$engine->utils->compareConstraints($tableName, $xmlCons, $dbCons, $deleteOnly);
     }
 
     /**
@@ -457,7 +446,7 @@ class DataBase
      */
     public function compareColumns($tableName, $xmlCols, $dbCols)
     {
-        return self::$utils->compareColumns($tableName, $xmlCols, $dbCols);
+        return self::$engine->utils->compareColumns($tableName, $xmlCols, $dbCols);
     }
 
     /**
@@ -490,6 +479,6 @@ class DataBase
      */
     public function sql2int($colName)
     {
-        return self::$engine->sql2int($colName);
+        return self::$engine->utilsSQL->sql2int($colName);
     }
 }
