+++ Core/Base/DataBase/DataBaseUtils.php
@@ -34,7 +34,7 @@ class DataBaseUtils
      * Enlace al motor de base de datos seleccionado en la configuración
      * @var DatabaseEngine
      */
-    private static $engine;
+    private $engine;
 
     /**
      * Construye y prepara la clase para su uso
@@ -42,7 +42,7 @@ class DataBaseUtils
      */
     public function __construct($engine)
     {
-        self::$engine = $engine;
+        $this->engine = &$engine;
     }
 
     /**
@@ -79,7 +79,7 @@ class DataBaseUtils
 
         $result = (
             (FS_CHECK_DB_TYPES !== '1') ||
-            self::$engine->compareDataTypes($db0, $xml) ||
+            $this->engine->compareDataTypes($db0, $xml) ||
             ($xml === 'serial') ||
             (
             strpos($db0, 'time') === 0 &&
@@ -111,20 +111,20 @@ class DataBaseUtils
 
             $column = $this->searchInArray($dbCols, 'name', $xml_col['nombre']);
             if (empty($column)) {
-                $result .= self::$engine->sqlAlterAddColumn($tableName, $xml_col);
+                $result .= $this->engine->utilsSQL->sqlAlterAddColumn($tableName, $xml_col);
                 continue;
             }
 
             if (!$this->compareDataTypes($column['type'], $xml_col['tipo'])) {
-                $result .= self::$engine->sqlAlterModifyColumn($tableName, $xml_col);
+                $result .= $this->engine->utilsSQL->sqlAlterModifyColumn($tableName, $xml_col);
             }
 
             if ($column['default'] === null && $xml_col['defecto'] !== '') {
-                $result .= self::$engine->sqlAlterConstraintDefault($tableName, $xml_col);
+                $result .= $this->engine->utilsSQL->sqlAlterConstraintDefault($tableName, $xml_col);
             }
 
             if ($column['is_nullable'] !== $xml_col['nulo']) {
-                $result .= self::$engine->sqlAlterConstraintNull($tableName, $xml_col);
+                $result .= $this->engine->utilsSQL->sqlAlterConstraintNull($tableName, $xml_col);
             }
         }
 
@@ -147,7 +147,7 @@ class DataBaseUtils
             if (strpos('PRIMARY;UNIQUE', $db_con['name']) === false) {
                 $column = $this->searchInArray($xmlCons, 'nombre', $db_con['name']);
                 if (empty($column)) {
-                    $result .= self::$engine->sqlDropConstraint($tableName, $db_con);
+                    $result .= $this->engine->utilsSQL->sqlDropConstraint($tableName, $db_con);
                 }
             }
         }
@@ -160,7 +160,7 @@ class DataBaseUtils
 
                 $column = $this->searchInArray($dbCons, 'name', $xml_con['nombre']);
                 if (empty($column)) {
-                    $result .= self::$engine->sqlAddConstraint($tableName, $xml_con['nombre'], $xml_con['consulta']);
+                    $result .= $this->engine->utilsSQL->sqlAddConstraint($tableName, $xml_con['nombre'], $xml_con['consulta']);
                 }
             }
         }
@@ -177,6 +177,6 @@ class DataBaseUtils
      */
     public function generateTable($tableName, $xmlCols, $xmlCons)
     {
-        return self::$engine->sqlCreateTable($tableName, $xmlCols, $xmlCons);
+        return $this->engine->utilsSQL->sqlCreateTable($tableName, $xmlCols, $xmlCons);
     }
 }
