+++ Core/Base/DataBase/DatabaseEngine.php
@@ -1,4 +1,5 @@
 <?php
+
 /**
  * This file is part of FacturaScripts
  * Copyright (C) 2015-2017  Carlos Garcia Gomez  carlos@facturascripts.com
@@ -19,7 +20,6 @@
 
 namespace FacturaScripts\Core\Base\DataBase;
 
-use mysqli;
 
 /**
  * Interface para cada uno de los motores de base de datos compatibles
@@ -29,13 +29,6 @@ use mysqli;
  */
 interface DatabaseEngine
 {
-
-    /**
-     * Genera el SQL para establecer las restricciones proporcionadas.
-     * @param array $xmlCons
-     */
-    public function generateTableConstraints($xmlCons);
-
     /**
      * Convierte los datos leidos del sqlColumns a estructura de trabajo
      * @param array $colData
@@ -135,12 +128,6 @@ interface DatabaseEngine
     public function dateStyle();
 
     /**
-     * Indica el SQL a usar para convertir la columna en Integer
-     * @param string $colName
-     */
-    public function sql2int($colName);
-
-    /**
      * Comprueba la existencia de una secuencia
      * @param mysqli|resource $link
      * @param string $tableName
@@ -156,90 +143,4 @@ interface DatabaseEngine
      * @param string $error
      */
     public function checkTableAux($link, $tableName, &$error);
-
-    /**
-     * Sentencia SQL para obtener el último valor de una secuencia o ID
-     */
-    public function sqlLastValue();
-
-    /**
-     * Sentencia SQL para obtener las columnas de una tabla
-     * @param string $tableName
-     */
-    public function sqlColumns($tableName);
-
-    /**
-     * Sentencia SQL para obtener las constraints de una tabla
-     * @param string $tableName
-     */
-    public function sqlConstraints($tableName);
-
-    /**
-     * Sentencia SQL para obtener las constraints (extendidas) de una tabla
-     * @param string $tableName
-     */
-    public function sqlConstraintsExtended($tableName);
-
-    /**
-     * Sentencia SQL para obtener los indices de una tabla
-     * @param string $tableName
-     */
-    public function sqlIndexes($tableName);
-
-    /**
-     * Sentencia SQL para crear una tabla
-     * @param string $tableName
-     * @param array $columns
-     * @param array $constraints
-     */
-    public function sqlCreateTable($tableName, $columns, $constraints);
-
-    /**
-     * Sentencia SQL para añadir una columna a una tabla
-     * @param string $tableName
-     * @param array $colData
-     */
-    public function sqlAlterAddColumn($tableName, $colData);
-
-    /**
-     * Sentencia SQL para modificar la definición de una columna de una tabla
-     * @param string $tableName
-     * @param array $colData
-     */
-    public function sqlAlterModifyColumn($tableName, $colData);
-
-    /**
-     * Sentencia SQL para modificar valor por defecto de una columna de una tabla
-     * @param string $tableName
-     * @param array $colData
-     */
-    public function sqlAlterConstraintDefault($tableName, $colData);
-
-    /**
-     * Sentencia SQL para modificar un constraint null de una columna de una tabla
-     * @param string $tableName
-     * @param array $colData
-     */
-    public function sqlAlterConstraintNull($tableName, $colData);
-
-    /**
-     * Sentencia SQL para eliminar una constraint de una tabla
-     * @param string $tableName
-     * @param array $colData
-     */
-    public function sqlDropConstraint($tableName, $colData);
-
-    /**
-     * Sentencia SQL para añadir una constraint de una tabla
-     * @param string $tableName
-     * @param string $constraintName
-     * @param string $sql
-     */
-    public function sqlAddConstraint($tableName, $constraintName, $sql);
-
-    /**
-     * Sentencia para crear una secuencia
-     * @param string $seqName
-     */
-    public function sqlSequenceExists($seqName);
 }
