+++ Core/Base/DataBase/Mysql.php
@@ -1,4 +1,5 @@
 <?php
+
 /**
  * This file is part of FacturaScripts
  * Copyright (C) 2013-2017  Carlos Garcia Gomez  carlos@facturascripts.com
@@ -19,6 +20,8 @@
 
 namespace FacturaScripts\Core\Base\DataBase;
 
+use FacturaScripts\Core\Base\DataBase\MysqlSQL;
+use FacturaScripts\Core\Base\DataBase\DataBaseUtils;
 use Exception;
 use mysqli;
 
@@ -32,17 +35,38 @@ class Mysql implements DatabaseEngine
 {
 
     /**
+     * El enlace con las utilidades comunes entre motores de base de datos.
+     * @var DataBaseUtils
+     */
+    public $utils;
+
+    /**
+     * Enlace al conjunto de sentencias SQL de la base de datos conectada
+     * @var DatabaseSQL; 
+     */
+    public $utilsSQL;
+    
+    /**
      * Relacion de Transacciones abiertas.
      * @var array
      */
     private $transactions;
 
     /**
+     * Ultimo mensaje de error
+     * @var string 
+     */
+    private $lastErrorMsg;
+    
+    /**
      * Contructor e inicializador de la clase
      */
     public function __construct()
     {
+        $this->utils = new DataBaseUtils($this);
+        $this->utilsSQL = new MysqlSQL($this);
         $this->transactions = [];
+        $this->lastErrorMsg = '';
     }
 
     /**
@@ -101,9 +125,10 @@ class Mysql implements DatabaseEngine
             return null;
         }
 
-        $result = @new \mysqli(FS_DB_HOST, FS_DB_USER, FS_DB_PASS, FS_DB_NAME, (int) FS_DB_PORT);
+        $result = new \mysqli(FS_DB_HOST, FS_DB_USER, FS_DB_PASS, FS_DB_NAME, (int) FS_DB_PORT);
         if ($result->connect_errno) {
             $error = $result->connect_error;
+            $this->lastErrorMsg = $error;
             return null;
         }
 
@@ -136,7 +161,7 @@ class Mysql implements DatabaseEngine
      */
     public function errorMessage($link)
     {
-        return $link->error;
+        return ($link->error != '') ? $link->error : $this->lastErrorMsg;
     }
 
     /**
@@ -211,6 +236,7 @@ class Mysql implements DatabaseEngine
                 $aux->free();
             }
         } catch (Exception $e) {
+            $this->lastErrorMsg = $e;
             $result = [];
         }
 
@@ -234,6 +260,7 @@ class Mysql implements DatabaseEngine
             }
             $result = (!$link->errno);
         } catch (Exception $e) {
+            $this->lastErrorMsg = $e;
             $result = false;
         }
 
@@ -258,16 +285,6 @@ class Mysql implements DatabaseEngine
     public function dateStyle()
     {
         return 'Y-m-d';
-    }
-
-    /**
-     * Devuelve el SQL necesario para convertir la columna a entero.
-     * @param string $colName
-     * @return string
-     */
-    public function sql2int($colName)
-    {
-        return 'CAST(' . $colName . ' as UNSIGNED)';
     }
 
     /**
@@ -384,79 +401,6 @@ class Mysql implements DatabaseEngine
     }
 
     /**
-     * Elimina código problemático de postgresql.
-     * @param string $sql
-     * @return string
-     */
-    private function fixPostgresql($sql)
-    {
-        $search = ['::character varying', 'without time zone', 'now()', 'CURRENT_TIMESTAMP', 'CURRENT_DATE'];
-        $replace = ['', '', "'00:00'", "'" . date('Y-m-d') . " 00:00:00'", date("'Y-m-d'")];
-        return str_replace($search, $replace, $sql);
-    }
-
-    /**
-     * Genera el SQL para establecer las restricciones proporcionadas.
-     * @param array $xmlCons
-     * @return string
-     */
-    public function generateTableConstraints($xmlCons)
-    {
-        $sql = '';
-        foreach ($xmlCons as $res) {
-            $sql .= ', CONSTRAINT ' . $res['nombre'] . ' ' . $res['consulta'];
-        }
-
-        return $this->fixPostgresql($sql);
-    }
-
-    /**
-     * TODO
-     * @param array $colData
-     * @return string
-     */
-    private function getConstraints($colData)
-    {
-        $notNull = ($colData['nulo'] === 'NO');
-        $result = ' NULL';
-        if ($notNull) {
-            $result = ' NOT' . $result;
-        }
-
-        $defaultNull = ($colData['defecto'] === null);
-        if ($defaultNull && !$notNull) {
-            $result .= ' DEFAULT NULL';
-        } else {
-            if ($colData['defecto'] !== '') {
-                $result .= ' DEFAULT ' . $colData['defecto'];
-            }
-        }
-
-        return $result;
-    }
-
-    /**
-     * Genera el SQL con el tipo de campo y las constraints DEFAULT y NULL
-     * @param array $colData
-     * @return string
-     */
-    private function getTypeAndConstraints($colData)
-    {
-        $type = stripos('integer,serial', $colData['tipo']) === FALSE ? strtolower($colData['tipo']) : FS_DB_INTEGER;
-        switch (TRUE) {
-            case ($type == 'serial'):
-            case (stripos($colData['defecto'], 'nextval(') !== FALSE):
-                $contraints = ' NOT NULL AUTO_INCREMENT';
-                break;
-
-            default:
-                $contraints = $this->getConstraints($colData);
-                break;
-        }
-        return ' ' . $type . $contraints;
-    }
-
-    /**
      * Convierte los datos leidos del sqlColumns a estructura de trabajo
      * @param array $colData
      * @return array
@@ -470,205 +414,5 @@ class Mysql implements DatabaseEngine
         unset($result['null'], $result['field']);
 
         return $result;
-    }
-
-    /**
-     * Devuleve el SQL para averiguar
-     * el último ID asignado al hacer un INSERT
-     * en la base de datos.
-     * @return string
-     */
-    public function sqlLastValue()
-    {
-        return 'SELECT LAST_INSERT_ID() as num;';
-    }
-
-    /**
-     * Devuelve el SQL para averiguar
-     * la lista de las columnas de una tabla.
-     * @param string $tableName
-     * @return string
-     */
-    public function sqlColumns($tableName)
-    {
-        return 'SHOW COLUMNS FROM `' . $tableName . '`;';
-    }
-
-    /**
-     * Devuelve el SQL para averiguar
-     * la lista de restricciones de una tabla.
-     * @param string $tableName
-     * @return string
-     */
-    public function sqlConstraints($tableName)
-    {
-        $sql = 'SELECT CONSTRAINT_NAME as name, CONSTRAINT_TYPE as type'
-            . ' FROM information_schema.table_constraints '
-            . ' WHERE table_schema = schema()'
-            . " AND table_name = '" . $tableName . "';";
-        return $sql;
-    }
-
-    /**
-     * Devuelve el SQL para averiguar
-     * la lista de restricciones avanzadas de una tabla.
-     * @param string $tableName
-     * @return string
-     */
-    public function sqlConstraintsExtended($tableName)
-    {
-        $sql = 'SELECT t1.constraint_name as name,'
-            . ' t1.constraint_type as type,'
-            . ' t2.column_name,'
-            . ' t2.referenced_table_name AS foreign_table_name,'
-            . ' t2.referenced_column_name AS foreign_column_name,'
-            . ' t3.update_rule AS on_update,'
-            . ' t3.delete_rule AS on_delete'
-            . ' FROM information_schema.table_constraints t1'
-            . ' LEFT JOIN information_schema.key_column_usage t2'
-            . ' ON t1.table_schema = t2.table_schema'
-            . ' AND t1.table_name = t2.table_name'
-            . ' AND t1.constraint_name = t2.constraint_name'
-            . ' LEFT JOIN information_schema.referential_constraints t3'
-            . ' ON t3.constraint_schema = t1.table_schema'
-            . ' AND t3.constraint_name = t1.constraint_name'
-            . ' WHERE t1.table_schema = SCHEMA()'
-            . " AND t1.table_name = '" . $tableName . "'"
-            . ' ORDER BY type DESC, name ASC;';
-        return $sql;
-    }
-
-    /**
-     * Devuelve el SQL para averiguar
-     * la lista de indices de una tabla.
-     * @param string $tableName
-     * @return string
-     */
-    public function sqlIndexes($tableName)
-    {
-        return 'SHOW INDEXES FROM ' . $tableName . ';';
-    }
-
-    /**
-     * Devuelve la sentencia SQL necesaria para crear una tabla con la estructura proporcionada.
-     * @param string $tableName
-     * @param array $columns
-     * @param array $constraints
-     * @return string
-     */
-    public function sqlCreateTable($tableName, $columns, $constraints)
-    {
-        $fields = '';
-        foreach ($columns as $col) {
-            $fields .= ', `' . $col['nombre'] . '` ' . $this->getTypeAndConstraints($col);
-        }
-
-        $sql = $this->fixPostgresql(substr($fields, 2));
-        return 'CREATE TABLE ' . $tableName . ' (' . $sql
-            . $this->generateTableConstraints($constraints) . ') '
-            . 'ENGINE=InnoDB DEFAULT CHARSET=utf8 COLLATE=utf8_bin;';
-    }
-
-    /**
-     * Sentencia SQL para añadir una columna a una tabla
-     * @param string $tableName
-     * @param array $colData
-     * @return string
-     */
-    public function sqlAlterAddColumn($tableName, $colData)
-    {
-        $sql = 'ALTER TABLE ' . $tableName . ' ADD `' . $colData['nombre'] . '` '
-            . $this->getTypeAndConstraints($colData) . ';';
-
-        return $sql;
-    }
-
-    /**
-     * Sentencia SQL para modificar una columna de una tabla
-     * @param string $tableName
-     * @param array $colData
-     * @return string
-     */
-    public function sqlAlterModifyColumn($tableName, $colData)
-    {
-        $sql = 'ALTER TABLE ' . $tableName
-            . ' MODIFY `' . $colData['nombre'] . '` '
-            . $this->getTypeAndConstraints($colData) . ';';
-
-        return $this->fixPostgresql($sql);
-    }
-
-    /**
-     * Sentencia SQL para modificar una constraint de una tabla
-     * @param string $tableName
-     * @param array $colData
-     * @return string
-     */
-    public function sqlAlterConstraintDefault($tableName, $colData)
-    {
-        $result = '';
-        if ($colData['tipo'] != 'serial') {
-            $result = $this->sqlAlterModifyColumn($tableName, $colData);
-        }
-        return $result;
-    }
-
-    /**
-     * Sentencia SQL para modificar una constraint NULL de un campo de una tabla
-     * @param string $tableName
-     * @param array $colData
-     * @return string
-     */
-    public function sqlAlterConstraintNull($tableName, $colData)
-    {
-        return $this->sqlAlterModifyColumn($tableName, $colData);
-    }
-
-    /**
-     * Sentencia SQL para eliminar una constraint de una tabla
-     * @param string $tableName
-     * @param array $colData
-     * @return string
-     */
-    public function sqlDropConstraint($tableName, $colData)
-    {
-        $start = 'ALTER TABLE ' . $tableName . ' DROP';
-        switch ($colData['type']) {
-            case 'FOREIGN KEY':
-                $sql = $start . ' FOREIGN KEY ' . $colData['name'] . ';';
-                break;
-
-            case 'UNIQUE':
-                $sql = $start . ' INDEX ' . $colData['name'] . ';';
-                break;
-
-            default:
-                $sql = '';
-        }
-        return $sql;
-    }
-
-    /**
-     * Sentencia SQL para añadir una constraint a una tabla
-     * @param string $tableName
-     * @param string $constraintName
-     * @param string $sql
-     * @return string
-     */
-    public function sqlAddConstraint($tableName, $constraintName, $sql)
-    {
-        return 'ALTER TABLE ' . $tableName
-            . ' ADD CONSTRAINT ' . $constraintName . ' '
-            . $this->fixPostgresql($sql) . ';';
-    }
-
-    /**
-     * Sentencia SQL para comprobar una secuencia
-     * @param string $seqName
-     * @return string
-     */
-    public function sqlSequenceExists($seqName)
-    {
-        return '';
     }
 }
