+++ Core/Base/DataBase/Postgresql.php
@@ -31,6 +31,34 @@ class Postgresql implements DatabaseEngine
 {
 
     /**
+     * El enlace con las utilidades comunes entre motores de base de datos.
+     * @var DataBaseUtils
+     */
+    public $utils;
+
+    /**
+     * Enlace al conjunto de sentencias SQL de la base de datos conectada
+     * @var DatabaseSQL; 
+     */
+    public $utilsSQL;
+
+    /**
+     * Ultimo mensaje de error
+     * @var string 
+     */
+    private $lastErrorMsg;
+    
+    /**
+     * Contructor e inicializador de la clase
+     */
+    public function __construct()
+    {
+        $this->utils = new DataBaseUtils($this);
+        $this->utilsSQL = new PostgresqlSQL($this);
+        $this->lastErrorMsg = '';
+    }
+    
+    /**
      * Devuelve el motor de base de datos y la versión.
      * @param resource $link
      * @return string
@@ -98,7 +126,8 @@ class Postgresql implements DatabaseEngine
      */
     public function errorMessage($link)
     {
-        return pg_last_error($link);
+        $error = pg_last_error($link);
+        return ($error != '') ? $error : $this->lastErrorMsg;
     }
 
     /**
@@ -173,6 +202,7 @@ class Postgresql implements DatabaseEngine
                 pg_free_result($aux);
             }
         } catch (Exception $e) {
+            $this->lastErrorMsg = $e;
             $result = [];
         }
 
@@ -223,17 +253,6 @@ class Postgresql implements DatabaseEngine
     }
 
     /**
-     * Devuelve el SQL necesario para convertir
-     * la columna a entero.
-     * @param string $colName
-     * @return string
-     */
-    public function sql2int($colName)
-    {
-        return 'CAST(' . $colName . ' as INTEGER)';
-    }
-
-    /**
      * Compara los tipos de datos de una columna. Devuelve TRUE si son iguales.
      * @param string $dbType
      * @param string $xmlType
@@ -281,7 +300,7 @@ class Postgresql implements DatabaseEngine
     {
         $aux = explode("'", $default);
         if (count($aux) === 3) {
-            $data = $this->select($link, $this->sqlSequenceExists($aux[1]));
+            $data = $this->select($link, $this->utilsSQL->sqlSequenceExists($aux[1]));
             if (empty($data)) {             /// ¿Existe esa secuencia?
                 $data = $this->select($link, 'SELECT MAX(' . $colname . ')+1 as num FROM ' . $tableName . ';');
                 $this->exec($link, 'CREATE SEQUENCE ' . $aux[1] . ' START ' . $data[0]['num'] . ';');
@@ -299,244 +318,5 @@ class Postgresql implements DatabaseEngine
     public function checkTableAux($link, $tableName, &$error)
     {
         return true;
-    }
-
-    /**
-     * Genera el SQL para establecer las restricciones proporcionadas.
-     * @param array $xmlCons
-     * @return string
-     */
-    public function generateTableConstraints($xmlCons)
-    {
-        $sql = '';
-
-        foreach ($xmlCons as $res) {
-            $value = strtolower($res['consulta']);
-            if (false !== strpos($value, 'primary key')) {
-                $sql .= ', ' . $res['consulta'];
-                continue;
-            }
-
-            if (FS_FOREIGN_KEYS === '1' || 0 !== strpos($res['consulta'], 'FOREIGN KEY')) {
-                $sql .= ', CONSTRAINT ' . $res['nombre'] . ' ' . $res['consulta'];
-            }
-        }
-
-        return $sql;
-    }
-
-    /**
-     * Devuleve el SQL para averiguar
-     * el último ID asignado al hacer un INSERT
-     * en la base de datos.
-     * @return string
-     */
-    public function sqlLastValue()
-    {
-        return 'SELECT lastval() as num;';
-    }
-
-    /**
-     * Devuelve el SQL para averiguar
-     * la lista de las columnas de una tabla.
-     * @param string $tableName
-     * @return string
-     */
-    public function sqlColumns($tableName)
-    {
-        $sql = 'SELECT column_name as name, data_type as type,'
-            . 'character_maximum_length, column_default as default,'
-            . 'is_nullable'
-            . ' FROM information_schema.columns'
-            . " WHERE table_catalog = '" . FS_DB_NAME . "'"
-            . " AND table_name = '" . $tableName . "'"
-            . ' ORDER BY 1 ASC;';
-
-        return $sql;
-    }
-
-    /**
-     * Devuelve el SQL para averiguar
-     * la lista de restricciones de una tabla.
-     * @param string $tableName
-     * @return string
-     */
-    public function sqlConstraints($tableName)
-    {
-        $sql = 'SELECT tc.constraint_type as type, tc.constraint_name as name'
-            . ' FROM information_schema.table_constraints AS tc'
-            . " WHERE tc.table_name = '" . $tableName . "'"
-            . " AND tc.constraint_type IN ('PRIMARY KEY','FOREIGN KEY','UNIQUE')"
-            . ' ORDER BY 1 DESC, 2 ASC;';
-        return $sql;
-    }
-
-    /**
-     * Devuelve el SQL para averiguar
-     * la lista de restricciones avanzadas de una tabla.
-     * @param string $tableName
-     * @return string
-     */
-    public function sqlConstraintsExtended($tableName)
-    {
-        $sql = 'SELECT tc.constraint_type as type, tc.constraint_name as name,'
-            . 'kcu.column_name,'
-            . 'ccu.table_name AS foreign_table_name, ccu.column_name AS foreign_column_name,'
-            . 'rc.update_rule AS on_update, rc.delete_rule AS on_delete'
-            . ' FROM information_schema.table_constraints AS tc'
-            . ' LEFT JOIN information_schema.key_column_usage AS kcu'
-            . ' ON kcu.constraint_schema = tc.constraint_schema'
-            . ' AND kcu.constraint_catalog = tc.constraint_catalog'
-            . ' AND kcu.constraint_name = tc.constraint_name'
-            . ' LEFT JOIN information_schema.constraint_column_usage AS ccu'
-            . ' ON ccu.constraint_schema = tc.constraint_schema'
-            . ' AND ccu.constraint_catalog = tc.constraint_catalog'
-            . ' AND ccu.constraint_name = tc.constraint_name'
-            . ' AND ccu.column_name = kcu.column_name'
-            . ' LEFT JOIN information_schema.referential_constraints rc'
-            . ' ON rc.constraint_schema = tc.constraint_schema'
-            . ' AND rc.constraint_catalog = tc.constraint_catalog'
-            . ' AND rc.constraint_name = tc.constraint_name'
-            . " WHERE tc.table_name = '" . $tableName . "'"
-            . " AND tc.constraint_type IN ('PRIMARY KEY','FOREIGN KEY','UNIQUE')"
-            . ' ORDER BY 1 DESC, 2 ASC;';
-
-        return $sql;
-    }
-
-    /**
-     * Devuelve el SQL para averiguar
-     * la lista de indices de una tabla.
-     * @param string $tableName
-     * @return string
-     */
-    public function sqlIndexes($tableName)
-    {
-        return "SELECT indexname as Key_name FROM pg_indexes WHERE tablename = '" . $tableName . "';";
-    }
-
-    /**
-     * Devuelve la sentencia SQL necesaria para crear una tabla con la estructura proporcionada.
-     * @param string $tableName
-     * @param array $columns
-     * @param array $constraints
-     * @return string
-     */
-    public function sqlCreateTable($tableName, $columns, $constraints)
-    {
-        $serials = ['serial', 'bigserial'];
-        $fields = '';
-        foreach ($columns as $col) {
-            $fields .= ', ' . $col['nombre'] . ' ' . $col['tipo'];
-
-            if ($col['nulo'] === 'NO') {
-                $fields .= ' NOT NULL';
-            }
-
-            if (in_array($col['tipo'], $serials, false)) {
-                continue;
-            }
-
-            if ($col['defecto'] !== '') {
-                $fields .= ' DEFAULT ' . $col['defecto'];
-            }
-        }
-
-        $sql = 'CREATE TABLE ' . $tableName . ' (' . substr($fields, 2)
-            . $this->generateTableConstraints($constraints) . ');';
-        return $sql;
-    }
-
-    /**
-     * Sentencia SQL para añadir una columna a una tabla
-     * @param string $tableName
-     * @param array $colData
-     * @return string
-     */
-    public function sqlAlterAddColumn($tableName, $colData)
-    {
-        $sql = 'ALTER TABLE ' . $tableName
-            . ' ADD COLUMN ' . $colData['nombre'] . ' ' . $colData['tipo'];
-
-        if ($colData['defecto'] !== '') {
-            $sql .= ' DEFAULT ' . $colData['defecto'];
-        }
-
-        if ($colData['nulo'] === 'NO') {
-            $sql .= ' NOT NULL';
-        }
-
-        return $sql . ';';
-    }
-
-    /**
-     * Sentencia SQL para modificar una columna a una tabla
-     * @param string $tableName
-     * @param array $colData
-     * @return string
-     */
-    public function sqlAlterModifyColumn($tableName, $colData)
-    {
-        $sql = 'ALTER TABLE ' . $tableName
-            . ' ALTER COLUMN ' . $colData['nombre'] . ' TYPE ' . $colData['tipo'];
-        return $sql . ';';
-    }
-
-    /**
-     * Sentencia SQL para modificar un valor por defecto de un campo de una tabla
-     * @param string $tableName
-     * @param array $colData
-     * @return string
-     */
-    public function sqlAlterConstraintDefault($tableName, $colData)
-    {
-        $action = ($colData['defecto'] !== '') ? ' SET DEFAULT ' . $colData['defecto'] : ' DROP DEFAULT';
-
-        return 'ALTER TABLE ' . $tableName . ' ALTER COLUMN ' . $colData['nombre'] . $action . ';';
-    }
-
-    /**
-     * Sentencia SQL para modificar una constraint null de un campo de una tabla
-     * @param string $tableName
-     * @param array $colData
-     * @return string
-     */
-    public function sqlAlterConstraintNull($tableName, $colData)
-    {
-        $action = ($colData['nulo'] === 'YES') ? ' DROP ' : ' SET ';
-        return 'ALTER TABLE ' . $tableName . ' ALTER COLUMN ' . $colData['nombre'] . $action . 'NOT NULL;';
-    }
-
-    /**
-     * Sentencia SQL para eliminar una constraint a una tabla
-     * @param string $tableName
-     * @param array $colData
-     * @return string
-     */
-    public function sqlDropConstraint($tableName, $colData)
-    {
-        return 'ALTER TABLE ' . $tableName . ' DROP CONSTRAINT ' . $colData['name'] . ';';
-    }
-
-    /**
-     * Sentencia SQL para añadir una constraint a una tabla
-     * @param string $tableName
-     * @param string $constraintName
-     * @param string $sql
-     * @return string
-     */
-    public function sqlAddConstraint($tableName, $constraintName, $sql)
-    {
-        return 'ALTER TABLE ' . $tableName . ' ADD CONSTRAINT ' . $constraintName . ' ' . $sql . ';';
-    }
-
-    /**
-     * Sentencia SQL para comprobar una secuencia
-     * @param string $seqName
-     * @return string
-     */
-    public function sqlSequenceExists($seqName)
-    {
-        return "SELECT * FROM pg_class where relname = '" . $seqName . "';";
     }
 }
