+++ Core/Base/MenuManager.php
@@ -20,7 +20,7 @@
 
 namespace FacturaScripts\Core\Base;
 
-use FacturaScripts\Core\Model;
+use FacturaScripts\Core\Base\Model;
 
 /**
  * Gestiona el uso del menú de Facturascripts
@@ -47,8 +47,7 @@ class MenuManager {
      */
     public function __construct($user) {
         if (!isset(self::$menu) || (self::$user !== $user)) {
-            self::$menu = $this->loadMenu($user);
-            self::$user = $user;
+            $this->getMenu($user);
         }        
     }
  
@@ -154,20 +153,5 @@ class MenuManager {
         }
         
         return self::$menu;
-    }
-    
-    /**
-     * Solo para pruebas. Imprime la estructura de menú
-     */
-    public function printMenu() {
-        foreach (self::$menu as $key => $value) {
-            print $value->title . " (" . $value->url . ")<br />";
-            foreach ($value->menu as $key2 => $value2) {
-                print "--->" . $value2->title . " (" . $value2->url . ")<br />";
-                foreach ($value2->menu as $key3 => $value3) {
-                    print "-------->" . $value3->title . " (" . $value3->url . ")<br />";
-                }
-            }
-        }        
-    }
+    }    
 }
